refactor: dto 패키지의 위치 ui -> application
