+++ src/main/java/kitchenpos/application/dto/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.application.dto;
 
 public class MenuGroupCreateRequest {
     private Long id;
