+++ src/main/java/kitchenpos/application/dto/ProductCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.application.dto;
 
 import java.math.BigDecimal;
